feat: Completed work on handle (community, collection, item) views and communities list.
This is the inital implementation of these features using AngularJs 1.5 and component architecture. The necessary node controllers, models and routes have been added.

The app is largely built with AngularJs components. There is very little outside of components and there would be very little reason at this point not to fold the remainer into components. Will pause to consider whehter or not that's a foolish consistency.

Issues: AC-623