feat: Completed initial work on browse searches including some extensive refactoring of core modules.
Starting to build a toxonomy of the solr queries required by the application. A node utility method selects the solr query based on input parameters. New core components in angular are the QueryManger and Utils services.  Defined constants for use across angular. A separate node module for server-side constants is in use currently. A fair amount of refactoring elsewhere.

Issue: AC-623