feat: Completed work on the solr query taxonomy and added count to author facets.
The solr query taxonomy implemented as an object.  It provides 7 URL templates that can be used with string interpolation.  Created a utility method that uses node utils.format() to add input parameters to solr queries. The new queries haven't yet been implemented in the server app models or in the client context object.

Issues: AC-623