chore: Refactored packages and added inline documentation.
The refactoring involved creating a new directory and moving nodes javascript files, but I prefer to think of this as similar to creating a new java package.  Inline documentation is now written for most methods in the app.  Time to move on to Angular Material and web design.

Issues: AC-623