feat: Added new solr query behavior for colletion page sort options and changed pager behavior.
The solr queries for title sort and jumpto queries are now in sync, so that you jump to the right place in the index regardless of the current sort order.  Changing the sort order also removes the jumpto query and sets the offset back to 0.  This needs to be replicated for the other fields (date, author, subject).  Also, pager now adds more results to the result set rather than replaces the current set.  To get more results, the user clicks a button. From the standpoint of usability, I think this added it of intentional user interaction is prefreable to infinite scroll.

Issues: AC-615