feat: Added mdDialog for item view.
Using a dialog for the item view may be the best option.  I created a new item.js controller on the server side, new Express route, and new angularjs service.  The itemDetail component now includes an mdDialog directive.  Attempting to dely the dialog show using resolve (promise). This seems to be working, but still seeing an awkward jump in the size of the dialog box after content layout.

Issues: AC-651