feat: Completed inital work on the item view in mdDialog.
The layout still needs to be reviewed with an item that contains multiple files, and the data included with each bitstream may need to be expanded to include the file name, truncated in necessary. Using a min-fab with arrows to show and hide extra metadata.  Not entirely satisfied with the look, but it does communicate effectively.

Issues: AC-617