feat: Final (tentative) animations added to the item list views.
These animations might not be ideal.  A simple fade animation would seem to be a fallback.  I was unable to make this work even though it should be simple.  Something overriding the animated opacity transistion seems like the obvious problem, but I also noticed that ng-enter, ng-leave, etc. were not added as expected. Fade can be done if we need something more subtle, but will require a bit of sleuthing to implement.

Issues: AC-651