feat: Refined animations for subject and author views and added title and navigation to browse views.
This should complete work on the collection page and the browse views that are accessed when browsing the collection authors and subjects.  Browse views are used for subject (or authors) that have more than 10 items.  In the current application configuration, browse screens are reached only via the collection view, so back navigation can be handled using $window.history.back().  The state of the collection window is maintained in application context, so back navigation results in returning to the previous browse list.

Issues: AC-615