feat: Implemented advanced search (discovery) filter and login message.
Final layout work remains to be done for advanced search, but the ability to add filters for title, author, subject and issue date and the ability to use two search modes (contains and not contains) are completed.  Some final functional requirements may need to be handled when the shared search box updated to support global and advanced searches.  I also added an itemTotal count to the REST community query result.  This can be used to hide/show a login for more collections message.

Issues: AC-678, AC-616