feat: Completed work on basic search boxes and provided user feedback in the event of an expired DSpace session.
Handle requests now recognize when data was not returned successfully.  The user is provided with an opportunity to log in again (or for the first time).  It's of course possible that data was not returned for some other reason, so need to avoid a loop of possible login attempts.  At the moment, it's not obvious how to distinguish between error causes.

Issues: AC-623, AC-651