bug: Fixed login problem with discovery views.
Logins from discovery views were returning to the communities view.  I added a new route and controller for retrieving the community list that did not set the session url to the communities path.  Also corrected a type issue in the angular controller, a conditional that resulted in an unnecessary API call that overrode the session url.  Logins from a discovery search now return to the that search prior to login, with additional results made available by the authentication.

Issues: AC-678