feat: Item handle page updated to newest layout and parent collection link added; util method for resetting query state.
The item page for handle request lays out properly for single and multi-file entries.  Also added a call to the GetCollectionInfo service to retrieve the parent community.  The parent community name and link are added to the page title section.  In addition, I added a method to the angular utils service that is called by the handle, discover and browse controllers to clear search and filter terms.

Issues: AC-617, AC-623