chore: Gruntfile and package.json updated and development server task running.
I'm updating this project based on configuration from the yoeman angular-fullstack generator.  This is a mean stack generator that includes much that we do not need for this project. But the Gruntfile and dependencies are similar to this project's current configuration, and support for tests is more complete. Some of the new optimizations and configuration appear to be useful, and this project might be refactored to take advantage, or not.  It's worth a look.

Issues: AC-680