feat: Added automatic redirection to the authentication when user attempt to access restricted content without a DSpace token.
This is implemented for handle requests.  It could be implemented for other requests if needed. The handle component redirects to auth/login if DSpace did not return data for the handle query.  Currently, when a restircted collection cannot be accessed by the user, the REST API returns a 500 error with no useful description. Description would be helpful since without it we can only assume that error was produced by invalid credentials.

Issues: AC-698