feat: This is a large commit with updates to many areas of the application needed before release.
In addition to many housekeeping tasks and UI updates, the commit includes new modules and code updates that allow for an advanced search option that is distinct from general discovery queries. Also included in an upgrade to Angular Material 1.1.0-rc4 and css tweaks for some of the new directives.

Issues: AC-623 AC-651