feat: Updated solr queries to use new environment configuration setting.
When working in development, the solr path values in environment.js are set to localhost:1234.  When in production mode, the values are empty strings on the assumption that the nodejs middleware and solr will be running on the same host and relative paths will work.  Added new getSolrUrl() method to utils.js.  All of this makes it easier to work with differing solr test enviroments during development.