feat: Added the parent collection to the item dialog display; author list rather than just first author to item list; set defaults on filters.
This work was done in response to user feedback.  That parent collection was an obvious omission.  The author is now an ng-repeat.  The filters now default to 'title' for the field and 'contains' for the search mode.  This might be helpful to users, we'll see.  Also, changed the search box placeholder label to 'enter search text' to see if the verb makes the intent clearer.

This information was being retrieved in the solr query but not displayed in the item view.