bug: Updated community component and application context to properly set the dspace session status and apply it to the community component login message.
Previously, when the user was logged into DSpace the login component was still visible if there was a difference in total and received collections. It should only appear if the user is not logged in and the collection counts do not match.  Since session status is checked asynchronously it we need to add a watch for session status update. Some of the mechanics around this had been broken by encapsulating the context object so methods for reading and updating the session status needed to be added to the AppContext module.

Issues: AC-743