feat: Sort options in the collection view are now using $location.search to maintain browser history.
This is a significant change that required introducing $scope watches for $locationChangeSuccess in pager and searchOptions.  With the exception of subject and author facet searches, sortOptions now sets the location query and lets pager respond to the location change by execing the query. If possible, it would make sense to move all queries to pager.  Init methods for sortOptions and pager are both checking the $location.search query string and initializing based on query status.

Issues: AC-729