bug: The item id is not being cleared from context at initialization.
Previously viewed items were loaded when a user leaves and returns to a collection.  The pager initialization has been updated to disable the item id when the component is intialized.  The query string is then checked and the contexted is updated with the new id one is provided.

Issue: AC-762