bug: The inline login component is incorrectly displaying in the community component.
Modified the community component controller so that the login component is set to visibile only when the watched dspace session status has changed.  Previously, we were updating without this check.  The new solution is better. It's unclear as to why this started to fail only after the recent updates to pager.

Issues: AC-763