bug: This fixes a few location change bugs and includes an seo test.
The bugs were fixed over the weekend.  Most notiable change was removing some unnecessary initialization in backPager.  Page should handle all initialization.  Also, it does not appear that google was following item links (ng-click) so I've added an href tag (href="#") to the element to see if that triggers generation of a dialog.

Issues: AC-761