feat: Completed work on latest Google indexing test with modified paging.
The application now uses links with href values for the main pager and for item titles.  Now testing to see how well Google indexes a public-facing journal.  History has been reworked again and initial tests show that it works well.  Some compromises were made in the case of filters for which reloading or backing through history to not work perfectly.  This should have no major effect on users, although filtering by subject and following a link to browse a large set of items does will return the user to the correct place in the index for the subject chosen but not to the identical place in the filtered set.  Since we do not include the back pager in filter results, this could be slightly disorienting.

Issue: AC-761