bug: Logos are not displaying in Internet Explorer.
The REST API is returning bitstreams with mime type 'application/octet-stream' and in the case of logos that will be displayed inline, this was causing Internet Explorer to abort the connection.  The current temporary solution is to check for the 'logo' file type in the bitstream module and set the mime type to 'image/jpg' in that case.  This is not ideal.  It would be much better to set the mime type based on the actual image logo mime type or extension.  This info is not provided by the REST API, at least as we are currently using it.  Resolving the issue for now.  Will add an improvement request to the backlog.

Issues: AC-752