feat: Added message screen for users who are directed to a collection without permission to access it.
The message appears when the user has a dspace session but the query for data failed to return.  The reason for failure might be something other than an authorization issue.  So the message content is configured in Messages and is easily configured as we learn more.  While I was at it, I also adjusted pager so that the button does not appear at the esnd of a set.  This was working earlier but was tweaked in a recent commit.

Issue: AC-775