feat: Completed adding the error message.
The error message now appears when an authenticated user cannot access an item.  Also, while working on this problem I added a new check on the community page. Authenticated users now do not see the login component simply because the count of collections shows that some are hidden from their view.  The login component only appears when the user is not logged in and collections are hidden from them.

Issue: AC-775