feat: Changed md-list-item to use ng-href instead of ng-click.
This is an attempt to make the application as easy as possible for Google to crawl.  Look and functionality are unchaged.  I also created a second inline item component for use with subject and author views.  The inline item component uses (and needs to use) ng-click and the controller's showItem() method rather than $location routing.

Issue: AC-761