bug: ng-repeat animation does not always complete.
After chaning the containing directive to ng-show (instead of ng-if) I saw consistent rather than intermittent animation failure on load.  By setting ready to be true in the init method, the animation was consistently successful.  The intent is to let the resetListView method toggle the ready to false and update methods to toogle to true when data is returned.  By setting ready to true on init one penalty is that the spinner is hidden.  If that's a problem we should probably bind the spinner ng-hide directive to a different variable.

Issues: AC-784