feat: The application should now work with REST running over ssl.
All of the connectors now have the additional option of 'rejectUnauthorized: false'. This prevents problems with self-signed certificates.  If or when we have a signed certificate on DSpace, this option is unnecessary and best removed for better security.

The dspace.js configuration now has a separate protocol for solr, so solr can continue to run over http.  The getBitstream module uses the node http module. For ssl, we need to import the 'https' module instead.

Issues: AC-586