feat: Added json-ld to the dialog item.
Added item type return element to the items.js model.  Value is from the dc.type value if provided in the item metadata. Did not include a default value, so if dc.type is not available, the json-ld @type will not be present.  That may not be valid json-ld.

Issues: AC-761