bug: Fixed item handle request.
The json-ld type field was overwriting the type returned by the dspace handle request.  When retrieveing indvidual items by handle, this field is used by the angular client to select the component.  I added a new field called jsonLdType to the response that is returned to the client.  The client handle component now has the handle item type and can display the component for single item results. The json-ld type is populated using the new jsonLdType response field.