feat: Added mobile-friendly (google) item views that do not use dialog that can be interpreted as an interstitial ad.
The collection component now recognizes that it is being initialized in mobile view by a request for an item.  Rather than load the list view and invoke the dialog, it retrieves the item data and displays inline. All possible user interactions (other than reloading the page) will continue to use the modal dialog. This is experimental. It assumes that search engines will make fresh requests for the item, forcing the app to initialize. This appears to be how google crawls javacript SPAs, but needs verification. All user links to the item from google results or from elsewhere invoke the new sans-dialog item view.

Issues: DSPACE-132