feat: Added prev to the google paging header link elements.
Following google's recommended practice, now including both next and prev links in the header.  These change to nofollow links as needed: at end of result set for next and at beginning for prev.  The href includes the correct offset values for next/prev and is empty for nofollow.

Issue: AC-761