feat: Changed the search engine paging list to use the item handle.
In this experiment, we're attempting to tell google and other engines to use only the paging links in the header (by adding nofollow rel attribute to the list item). Then, in the paging componenent, the item detail uses the actual item handle rather than the url for the item within the main application context.  The goal is to make search engine indexing more transparent and improve search result listings, while at the same time allow crawling of live pages. Although difficult to interpret given what I see in current google search results, it appears that all is not well with google's ability to catch page title updates, etc. Directing the search engine to a standalone page for the item may make more sense to the user as well.

Issues: AC-761