bug: Using the browser back button from a discovery/advanced search page to return to collection page with item list was not returning items.
This only happens when the collection page is initially opened with no query parameters. The cause was a check in PagerUtils.setQueryComponents that used the query type and took no action of the query type was DISCOVER.  Since the user had just exectuted a DISCOVER search, this had the effect of not setting the query back to default values.  Now using the pager's context attribute for the commparison.  This is a QueryAction constant that is passed in from the parent component.  It cannot be overriden by query object updates.

Issues: DSPACE-137