feat: Another experiment with optimizing for google crawler.
The only path to paging and viewing items is now though using the 'next' and 'prev' links provided in the header.  All other paging and item links are set to 'nofollow', the search engine links are now generated by a separate service 'SeoPaging' and the node solr query service returns 404 when items are not found.

Issues: AC-761