feat: Updates related to googlebot crawling (other search engines still untested)
Many of the updates are optimizations to assure that title and metadata values are correct as one navigates the application. Now handling 404 errors correctly for item and facet lookups.  Also added a utility method for detecting search engine user agents.  The collection component is now using this as context, prompting the item-list-component to use the more canonical item handle url for links.  If not a search engine user agent, the item list still uses the modal dialog to present item information.

Issues: AC-761