chore: Removed more dependencies on global $scope.
The loader component now provides all logic related to initialization and location changes.  It's implemented as a child of the itemListComponent and is a sibling of pagers, sort options and detail views.  The loader component has been tested with all components and is working correctly for the most part.  Still a few issues with the counts displayed in the view for some sort and paging operations.