feat: Added switch for search engine crawlers and completed more layout work.
Attempting to use Express routing to switch between item views based on useragent.  The goal is to direct search engines to the canonical handle url for items.  This was not possible in client-side angular modules, but should be possible server-side.  Although google is successfully crawling the modal dialog item views that we use in the application, this isn't optimal and could lead to temporarily broken links in collections that are updated frequently.  This commit also includes a few additional view updates for mobile and adjustments to index.html and paritals.

Issues: AC-761