chore: Reorganized the public app module and template layout.
Templates now located in the public 'ds' directory and can therefore be loaded using the Express static middleware.  This removes a lot of boilerplate routing from routes.js.  Services are now in a separate directory and organized in new sub-directories.  Only component files live in the component directory. This structure should make the entire app easier to navigate for the developer.