chore: Modified routes to support search engine view.
The express route designed to return a link to the canonical url for search engines was not working with static routes.  The static route was overriding the route middleware for alternate urls.  Since I want to keep static routing first in configuration, decided to move the templates out of the static route directory and into an alternate path.  The middleware for alternate routes now handling requests for the itemDetail template.

Issues: AC-761