chore: Continued restructing of the application and modified several views.
This commit includes the addition of a new angular application configuration option for the root path 'ds.' It's designed to make it easier to modify the url path by passing the configured parameter into all routes and component template paths.  The root paths for the application are now: 'ds' for partials and index.html, 'ds-app' for all application javascript, css and images, and 'ds-api' for xhr requests. This commit should also complete work on top-level layouts.  It also addresses a feature request for the author to appear in discovery search results.