feat: The discovery results now use item-level authentication.
The best solution is to add a login button to the dialog.  Still need to add switch in the dialog that notifies authenticated users that they do not have access to item.  The discovery search forms need to be updated to work with new design.  Once these steps are finished may be able to resolve the issue.

Issues: AC-951