fix: The mobile page for after successful login now displays data.
The issue here may simply be initializing requests prior to the onInit hook method call.  When this application was first written the onInit method either did not exist or was undocumented.  With this more recent build of the application is using angular 1.6.4 rather than the earlier 1.5 version. Should probably shrinkwrap.

Issues: DSPACE-178