bugfix: The auto login option for solr discovery queries is now working.
The problem was caused by a bad url path entered into Tagger.  Discovery routes defined for angualar can take an auth parameter.  This paremeter cannot
be used in combiniation with auto login.  It fact, it' not clear that this auth route parameter is necessary.  I've left it in place and added a TODO
for later review. Given that there were a few other minor changes to the code base, I've compiled and deployed the new code as a minor release (v1.2.3).

Issues: AC-1186